2026.7.11 拆分sensor为status和actuator两个独立节点
- sensor_status_node.py: 纯STATUS帧接收，发布/sensor_status（独立TCP连接）
- sensor_actuator_node.py: 执行器控制+ACK/ACTUATOR_FB反馈，订阅/auv_actuator_control，发布/auv_actuator_status
- 两个节点各自建立独立TCP连接，互不耦合
- 更新 begin.launch 和 data_collection.launch 引用新节点
