+++ src/auv_control/launch/begin.launch
@@ -5,14 +5,19 @@
 
     
     <!-- <node pkg="auv_control" type="main_driver.py" name="main_driver" output="screen" /> -->
-    <node pkg="auv_control" type="sensor_driver_v2.py" name="sensor_driver" output="screen" >
+    <!-- sensor STATUS 接收节点（纯上行，不发送控制帧） -->
+    <node pkg="auv_control" type="sensor_status_node.py" name="sensor_status_node" output="screen" >
         <param name="sensor_ip" value="$(arg vehicle_ip)" />
         <param name="sensor_port" value="5064" />
-        <param name="light1_brightness" value="0" /> 
-        <param name="light2_brightness" value="0" />
         <param name="save_raw_data" type="bool" value="true" />
         <param name="raw_save_dir" type="string" value="$(arg save_dir)" />
         <param name="raw_flush_every" type="int" value="1" />
+    </node>
+
+    <!-- sensor 执行器控制+反馈节点（下行命令 + ACK/ACTUATOR_FB 接收） -->
+    <node pkg="auv_control" type="sensor_actuator_node.py" name="sensor_actuator_node" output="screen" >
+        <param name="sensor_ip" value="$(arg vehicle_ip)" />
+        <param name="sensor_port" value="5064" />
     </node>
 
     <!-- <node pkg="auv_control" type="state_control.py" name="state_control" output="screen" /> -->
