+++ src/auv_control/scripts/sensor_actuator_node.py
@@ -1,0 +1,437 @@
+#! /usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+"""
+名称：sensor_actuator_node.py
+功能：执行器控制与反馈节点，通过 sensor debug 协议下行控制执行机构，
+     接收 ACK 和 ACTUATOR_FB 上行帧并发布执行机构实际状态
+作者：buyegaid
+监听：/auv_actuator_control (ActuatorControl.msg)
+发布：/auv_actuator_status (ActuatorControl.msg)
+记录：
+2026.7.11
+    从 sensor_driver_v2.py 拆分出执行器控制逻辑，独立 TCP 连接
+    发布 /auv_actuator_status 反馈执行机构实际状态
+"""
+
+import socket
+import threading
+
+import rospy
+
+from auv_control.msg import ActuatorControl
+
+
+class SensorActuatorNode:
+    """
+    sensor 执行器控制与反馈节点：
+    - 独立 TCP 连接到 sensor:5064
+    - 订阅 /auv_actuator_control，下发 CAMERA_LIGHT_SET + ACTUATOR_SET
+    - 接收 ACK 和 ACTUATOR_FB 上行帧
+    - 发布 /auv_actuator_status 反馈执行机构实际状态
+    - 忽略 STATUS 周期帧（由 sensor_status_node 处理）
+    """
+
+    # --- 上行帧常量 ---
+    PACKET_LEN = 64
+    HEADER = b'\xFE\xEF'
+    TAIL = b'\xFA\xAF'
+    REPORT_ACK = 0x01
+    REPORT_ACTUATOR_FB = 0x03
+
+    # --- 下行帧常量 (debug 协议) ---
+    DOWNLINK_LEN = 54
+    DOWNLINK_HEADER = b'\xFE\xFE'
+    DOWNLINK_TAIL = b'\xFD\xFD'
+    PROTOCOL_VERSION = 0x02
+
+    CMD_CAMERA_LIGHT = 0x10
+    CMD_ACTUATOR = 0x30
+    OP_SET = 0x00
+    FLAG_NEED_ACK = 0x01
+    SEND_RATE_HZ = 5
+
+    def __init__(self):
+        self.ip = rospy.get_param('~sensor_ip', '192.168.1.115')
+        self.port = rospy.get_param('~sensor_port', 5064)
+
+        self.sock = None
+        self.buffer = bytearray()
+
+        # --- 执行器状态缓存 ---
+        self.lock = threading.Lock()
+        self.seq = 0  # 下行帧序列号 (0-255)
+
+        self.light1 = 0
+        self.light2 = 0
+        self.heading_servo = 0x80
+        self.clamp_servo = 0xFF
+        self.drive_cmd = 0
+        self.drive_speed = 0
+        self.red_light = 0
+        self.yellow_light = 0
+        self.green_light = 0
+        self.control_changed = False
+
+        # --- 执行机构实际反馈缓存（来自 ACTUATOR_FB）---
+        self.fb_heading = 0x80
+        self.fb_clamp = 0xFF
+        self.fb_drive_cmd = 0
+        self.fb_drive_speed = 0
+        self.fb_red = 0
+        self.fb_yellow = 0
+        self.fb_green = 0
+
+        # --- 发送线程 ---
+        self.is_sending = True
+        self.send_thread = None
+
+        # --- ROS 接口 ---
+        rospy.Subscriber('/auv_actuator_control', ActuatorControl, self.actuator_callback)
+        self.status_pub = rospy.Publisher('/auv_actuator_status', ActuatorControl, queue_size=10)
+
+        self.connect()
+        rospy.loginfo("sensor_actuator_node: 已启动（执行器控制+反馈模式）")
+
+    # ============================================================
+    # TCP 连接
+    # ============================================================
+
+    def connect(self):
+        while not rospy.is_shutdown():
+            try:
+                self.sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+                self.sock.connect((self.ip, self.port))
+                self.sock.settimeout(1.0)
+                rospy.loginfo(f"sensor_actuator_node: TCP连接 {self.ip}:{self.port}")
+                return
+            except Exception as e:
+                rospy.logerr(f"sensor_actuator_node: TCP连接失败 {e}, 2s 后重试")
+                self.sock = None
+                rospy.sleep(2)
+
+    # ============================================================
+    # 上行帧校验
+    # ============================================================
+
+    @staticmethod
+    def calc_xor(packet):
+        """上行帧 XOR 校验：覆盖字节 0-60"""
+        value = 0
+        for byte in packet[0:61]:
+            value ^= byte
+        return value
+
+    def verify_packet(self, packet):
+        if len(packet) != self.PACKET_LEN:
+            return False
+        if packet[0:2] != self.HEADER:
+            return False
+        if packet[62:64] != self.TAIL:
+            return False
+        return self.calc_xor(packet) == packet[61]
+
+    # ============================================================
+    # 上行接收与同步
+    # ============================================================
+
+    def recv_loop(self):
+        while not rospy.is_shutdown():
+            try:
+                if self.sock is None:
+                    self.connect()
+                    continue
+
+                data = self.sock.recv(1024)
+                if not data:
+                    raise RuntimeError("对端关闭连接")
+
+                self.buffer.extend(data)
+                self._process_buffer()
+
+            except Exception as e:
+                rospy.logerr(f"sensor_actuator_node: 接收失败: {e}")
+                try:
+                    if self.sock:
+                        self.sock.close()
+                except Exception:
+                    pass
+                self.sock = None
+                self.buffer = bytearray()
+                rospy.sleep(1)
+
+    def _process_buffer(self):
+        while len(self.buffer) >= self.PACKET_LEN:
+            idx = self.buffer.find(self.HEADER)
+            if idx < 0:
+                if len(self.buffer) > 1:
+                    self.buffer = self.buffer[-1:]
+                return
+
+            if idx > 0:
+                rospy.logdebug(f"sensor_actuator_node: 跳过 {idx} 字节同步到帧头")
+                del self.buffer[:idx]
+
+            if len(self.buffer) < self.PACKET_LEN:
+                return
+
+            packet = bytes(self.buffer[:self.PACKET_LEN])
+            checksum_ok = self.verify_packet(packet)
+            tail_ok = (packet[62:64] == self.TAIL)
+
+            if checksum_ok:
+                self._parse(packet)
+                del self.buffer[:self.PACKET_LEN]
+            else:
+                if not tail_ok:
+                    # 帧尾不匹配 → 假帧头，跳到下一处 FE EF
+                    next_idx = self.buffer.find(self.HEADER, 2)
+                    if next_idx > 0:
+                        rospy.logdebug(
+                            f"sensor_actuator_node: 假帧头（帧尾不匹配），跳过 {next_idx} 字节到下一帧头"
+                        )
+                        del self.buffer[:next_idx]
+                    else:
+                        del self.buffer[0]
+                else:
+                    rospy.logwarn("sensor_actuator_node: 报文校验失败(XOR错误)，丢弃1字节继续同步")
+                    del self.buffer[0]
+
+    # ============================================================
+    # 上行帧解析
+    # ============================================================
+
+    def _parse(self, packet):
+        """解析 ACK 和 ACTUATOR_FB 帧，忽略 STATUS 帧"""
+        report_type = packet[4]
+
+        if report_type == self.REPORT_ACK:
+            self._parse_ack(packet)
+        elif report_type == self.REPORT_ACTUATOR_FB:
+            payload_len = packet[9]
+            self._parse_actuator_fb(packet, payload_len)
+        # STATUS (0x00) 和 CONFIG_FB (0x02) 由 status 节点处理，本节点忽略
+
+    def _parse_ack(self, packet):
+        """解析 ACK 帧，记录命令处理结果"""
+        ack_cmd = packet[5]
+        ack_op = packet[6]
+        ack_result = packet[7]
+        ack_error = packet[8]
+
+        cmd_name = {
+            self.CMD_CAMERA_LIGHT: 'CAMERA_LIGHT',
+            self.CMD_ACTUATOR: 'ACTUATOR',
+        }.get(ack_cmd, f'0x{ack_cmd:02X}')
+
+        if ack_result != 0:
+            rospy.logwarn(
+                f"sensor_actuator_node: ACK错误 cmd={cmd_name} op=0x{ack_op:02X} "
+                f"result=0x{ack_result:02X} error=0x{ack_error:02X}"
+            )
+        else:
+            rospy.loginfo(f"sensor_actuator_node: ACK成功 cmd={cmd_name}")
+
+    def _parse_actuator_fb(self, packet, payload_len):
+        """解析 ACTUATOR_FB 帧，更新反馈缓存并发布 /auv_actuator_status"""
+        if payload_len < 8:
+            rospy.logwarn(f"sensor_actuator_node: ACTUATOR_FB payload长度不足: {payload_len}")
+            return
+
+        payload = packet[10:10 + payload_len]
+        fb_result = packet[7]
+        fb_error = packet[8]
+
+        # 更新反馈缓存
+        with self.lock:
+            self.fb_heading = payload[0]
+            self.fb_clamp = payload[1]
+            self.fb_drive_cmd = payload[2]
+            self.fb_drive_speed = payload[3]
+            self.fb_red = payload[4]
+            self.fb_yellow = payload[5]
+            self.fb_green = payload[6]
+
+        rospy.loginfo(
+            f"sensor_actuator_node: 执行机构反馈 "
+            f"heading=0x{payload[0]:02X} angle=0x{payload[1]:02X} "
+            f"drive=({payload[2]},{payload[3]}) "
+            f"led=({payload[4]},{payload[5]},{payload[6]}) "
+            f"result=0x{fb_result:02X} error=0x{fb_error:02X}"
+        )
+
+        # 发布 /auv_actuator_status
+        self._publish_status()
+
+    def _publish_status(self):
+        """根据反馈缓存和命令缓存组装 ActuatorControl 并发布"""
+        with self.lock:
+            msg = ActuatorControl()
+            # 补光灯：来自最后命令值（CAMERA_LIGHT 无硬件反馈）
+            msg.light1 = self.light1
+            msg.light2 = self.light2
+            # 执行机构：来自 ACTUATOR_FB 硬件反馈
+            msg.heading_servo = self.fb_heading
+            msg.clamp_servo = self.fb_clamp
+            msg.drive_cmd = self.fb_drive_cmd
+            msg.drive_speed = self.fb_drive_speed
+            msg.red_light = self.fb_red
+            msg.yellow_light = self.fb_yellow
+            msg.green_light = self.fb_green
+
+        self.status_pub.publish(msg)
+
+    # ============================================================
+    # 下行控制
+    # ============================================================
+
+    def actuator_callback(self, msg):
+        """接收 /auv_actuator_control 控制指令，更新缓存并置脏标志"""
+        try:
+            new_light1 = max(0, min(100, msg.light1))
+            new_light2 = max(0, min(100, msg.light2))
+            new_heading = max(0, min(255, msg.heading_servo))
+            new_clamp = max(0, min(255, msg.clamp_servo))
+            new_drive_cmd = msg.drive_cmd if msg.drive_cmd in (0, 1, 2) else 0
+            new_drive_speed = max(0, min(254, msg.drive_speed))
+            new_red = 1 if msg.red_light else 0
+            new_yellow = 1 if msg.yellow_light else 0
+            new_green = 1 if msg.green_light else 0
+
+            changed = (
+                new_light1 != self.light1 or
+                new_light2 != self.light2 or
+                new_heading != self.heading_servo or
+                new_clamp != self.clamp_servo or
+                new_drive_cmd != self.drive_cmd or
+                new_drive_speed != self.drive_speed or
+                new_red != self.red_light or
+                new_yellow != self.yellow_light or
+                new_green != self.green_light
+            )
+
+            with self.lock:
+                self.light1 = new_light1
+                self.light2 = new_light2
+                self.heading_servo = new_heading
+                self.clamp_servo = new_clamp
+                self.drive_cmd = new_drive_cmd
+                self.drive_speed = new_drive_speed
+                self.red_light = new_red
+                self.yellow_light = new_yellow
+                self.green_light = new_green
+                if changed:
+                    self.control_changed = True
+
+            if changed:
+                rospy.loginfo(
+                    f"sensor_actuator_node: 执行器控制更新 "
+                    f"light=({self.light1},{self.light2}) "
+                    f"heading=0x{self.heading_servo:02X} clamp=0x{self.clamp_servo:02X} "
+                    f"drive=({self.drive_cmd},{self.drive_speed}) "
+                    f"led=({self.red_light},{self.yellow_light},{self.green_light})"
+                )
+        except Exception as e:
+            rospy.logerr(f"sensor_actuator_node: 执行器控制回调失败: {e}")
+
+    def _next_seq(self):
+        """递增并返回序列号 (0-255 回绕)"""
+        self.seq = (self.seq + 1) & 0xFF
+        return self.seq
+
+    @staticmethod
+    def _calc_downlink_xor(packet):
+        """下行帧异或校验: 字节 0-50"""
+        xor = 0
+        for i in range(51):
+            xor ^= packet[i]
+        return xor & 0xFF
+
+    def build_camera_light_frame(self):
+        """构造 CAMERA_LIGHT_SET 下行帧 (cmd=0x10, op=0x00)"""
+        packet = bytearray(self.DOWNLINK_LEN)
+        packet[0:2] = self.DOWNLINK_HEADER
+        packet[2] = self.PROTOCOL_VERSION
+        packet[3] = self._next_seq()
+        packet[4] = self.CMD_CAMERA_LIGHT
+        packet[5] = self.OP_SET
+        packet[6] = 0x00
+        packet[7] = 2
+        packet[8] = self.light1
+        packet[9] = self.light2
+        packet[40] = self.FLAG_NEED_ACK
+        packet[51] = self._calc_downlink_xor(packet)
+        packet[52:54] = self.DOWNLINK_TAIL
+        return packet
+
+    def build_actuator_frame(self):
+        """构造 ACTUATOR_SET 下行帧 (cmd=0x30, op=0x00)"""
+        packet = bytearray(self.DOWNLINK_LEN)
+        packet[0:2] = self.DOWNLINK_HEADER
+        packet[2] = self.PROTOCOL_VERSION
+        packet[3] = self._next_seq()
+        packet[4] = self.CMD_ACTUATOR
+        packet[5] = self.OP_SET
+        packet[6] = 0x00
+        packet[7] = 7
+        packet[8] = self.heading_servo
+        packet[9] = self.clamp_servo
+        packet[10] = self.drive_cmd
+        packet[11] = self.drive_speed
+        packet[12] = self.red_light
+        packet[13] = self.yellow_light
+        packet[14] = self.green_light
+        packet[40] = self.FLAG_NEED_ACK
+        packet[51] = self._calc_downlink_xor(packet)
+        packet[52:54] = self.DOWNLINK_TAIL
+        return packet
+
+    def send_loop(self):
+        """5Hz 发送线程: 仅在控制值变化时发送两帧"""
+        rate = rospy.Rate(self.SEND_RATE_HZ)
+        while not rospy.is_shutdown() and self.is_sending:
+            try:
+                if self.sock is None:
+                    rate.sleep()
+                    continue
+
+                with self.lock:
+                    changed = self.control_changed
+                    if changed:
+                        self.control_changed = False
+
+                if changed:
+                    light_frame = self.build_camera_light_frame()
+                    self.sock.sendall(bytes(light_frame))
+                    actuator_frame = self.build_actuator_frame()
+                    self.sock.sendall(bytes(actuator_frame))
+                    rospy.loginfo("sensor_actuator_node: 已发送控制帧")
+            except Exception as e:
+                rospy.logerr(f"sensor_actuator_node: 发送失败: {e}")
+            rate.sleep()
+
+    # ============================================================
+
+    def spin(self):
+        self.send_thread = threading.Thread(target=self.send_loop, daemon=True)
+        self.send_thread.start()
+        try:
+            self.recv_loop()
+        finally:
+            self.is_sending = False
+            if self.send_thread and self.send_thread.is_alive():
+                self.send_thread.join(timeout=2)
+            if self.sock:
+                try:
+                    self.sock.close()
+                except Exception:
+                    pass
+
+
+if __name__ == "__main__":
+    rospy.init_node('sensor_actuator_node')
+    try:
+        node = SensorActuatorNode()
+        node.spin()
+    except rospy.ROSInterruptException:
+        pass
